liver_fibrosis_flask/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.2.3
tensorflow==2.12.0
pillow==10.2.0
